refactor(validate): extract per-node config validation into its own function
Move the large block of per-node config validation logic from `validate_all` into a dedicated `validate_node_configs` function, reducing the main validation function by over 300 lines and improving readability. The extracted function handles sub-workflow, fan-out, memory, dedup, and approval node config checks, keeping the same validation behaviour.

Auto-committed-on: dragonfly
